add useAttendantPushNotifications hook to handle frontend push notification logic

+++ frontend/src/hooks/useAttendantPushNotifications.ts
@@ -81,7 +81,7 @@ export function useAttendantPushNotifications(token: string): UseAttendantPushNo
       const reg = await navigator.serviceWorker.ready;
       const sub = await reg.pushManager.subscribe({
         userVisibleOnly:      true,
-        applicationServerKey: urlBase64ToUint8Array(publicKey) as Uint8Array,
+        applicationServerKey: urlBase64ToUint8Array(publicKey),
       });
 
       // 4. Extract keys as URL-safe base64
